+++ src/main/org/jboss/messaging/core/journal/SequentialFileFactory.java
@@ -34,7 +34,7 @@ import java.util.List;
  */
 public interface SequentialFileFactory
 {
-	SequentialFile createSequentialFile(String fileName, boolean sync, int maxIO, int timeout) throws Exception;
+	SequentialFile createSequentialFile(String fileName, boolean sync, int maxIO, long timeout) throws Exception;
 	
 	List<String> listFiles(String extension) throws Exception;
