+++ src/main/org/jboss/messaging/core/journal/impl/AIOSequentialFileFactory.java
@@ -26,7 +26,7 @@ public class AIOSequentialFileFactory extends AbstractSequentialFactory
 		super(journalDir);
 	}
 	
-	public SequentialFile createSequentialFile(final String fileName, final boolean sync, final int maxIO, final int timeout) throws Exception
+	public SequentialFile createSequentialFile(final String fileName, final boolean sync, final int maxIO, final long timeout) throws Exception
 	{
 		return new AIOSequentialFile(journalDir, fileName, maxIO, timeout);
 	}
